Issue #2091 - Cache tries to remove folders. Check with is_file() added.

+++ wp-content/mu-plugins/wp-really-simple-cache.php
@@ -314,7 +314,7 @@ if (!function_exists('WpSimpleCache_purge_post_by_id')) {
 				//Purge this post 
 				$file_name = $wp_simple_cache::get_cache_dir().$wp_simple_cache::get_filename_from_url(get_permalink( $post_id ));
 				
-				if ( file_exists( $file_name ) ) {
+				if ( file_exists( $file_name ) && is_file( $file_name ) ) {
 					unlink($file_name);
 				}
 				
@@ -323,7 +323,7 @@ if (!function_exists('WpSimpleCache_purge_post_by_id')) {
 					$post_parent_id = wp_get_post_parent_id( $post_id );  
 					if ( $post_parent_id !== 0 && is_numeric( $post_parent_id ) ) {
 						$file_name = $wp_simple_cache::get_cache_dir().$wp_simple_cache::get_filename_from_url(get_permalink( $post_parent_id ));
-						if ( file_exists( $file_name ) ) {
+						if ( file_exists( $file_name ) && is_file( $file_name ) ) {
 							unlink($file_name);
 						}				
 					}
@@ -331,7 +331,7 @@ if (!function_exists('WpSimpleCache_purge_post_by_id')) {
 				
 				//Purge archive page 
 				$file_name = $wp_simple_cache::get_cache_dir().$wp_simple_cache::get_filename_from_url(get_post_type_archive_link(get_post_type( $post_id ) ) ); 
-				if ( file_exists( $file_name ) ) {
+				if ( file_exists( $file_name ) && is_file( $file_name ) ) {
 					unlink($file_name);
 				}
